[IOV6] Added "simple pet AI" and AI for non-controlled dk ghoul.
Use this AI to add spells defined in the creature_template to the
pet autocast map (e.g. for guardian/protector pets). Ofc only possible
for spells with cooldown.
The AI used for any actions is the normal PetAI then.

The Ghoul AI is a bit hacky... but who cares ;-)

Signed-off-by: Janu <jaed1@gmx.net>